+++ solvers/__init__.py
@@ -1,0 +1,3 @@
+from .base import BaseWSPSolver
+from .ortools import SolutionCollector, ORToolsCSWSPSolver, ORToolsPBPBWSPSolver, ORToolsUDPBWSPSolver
+from .z3 import Z3UDPBWSPSolver, Z3PBPBWSPSolver